feat: 원본 기도를 지우면 모임 기도도 같이 삭제되게 변경
