+++ src/main/java/com/uspray/uspray/infrastructure/PrayRepository.java
@@ -6,6 +6,7 @@ import com.uspray.uspray.exception.model.NotFoundException;
 import com.uspray.uspray.infrastructure.querydsl.pray.PrayRepositoryCustom;
 import java.time.LocalDate;
 import java.util.List;
+import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -18,6 +19,7 @@ public interface PrayRepository extends JpaRepository<Pray, Long>, PrayRepositor
                 ErrorStatus.PRAY_NOT_FOUND_EXCEPTION.getMessage()));
     }
 
+    @EntityGraph(attributePaths = {"groupPray"})
     default Pray getPrayByIdAndMemberId(Long prayId, String username) throws NotFoundException {
         return findById(prayId)
             .filter(pray -> pray.getMember().getUserId().equals(username))
